+++ lib/src/core/helper/utils.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+class HelperUtils {
+  static String getHomeDirectory() {
+    if (Platform.isWindows) {
+      return Platform.environment['USERPROFILE']!;
+    } else {
+      return Platform.environment['HOME']!;
+    }
+  }
+}