+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- configuration system support
+- complete theme system with dark/light mode
+
 ## [0.1.0] - 2026-02-13
 
 ### Added